Moves repositories to the domain layer
Moves the repository interfaces from the data layer to the domain layer.
This change promotes a cleaner architecture by ensuring that the domain layer,
which contains the core business logic, is independent of the data layer.
It also prepares the ground for implementing use cases and keeps a clean architecture.